competitors api update
