+++ tournament-manager-be/Tournament.Manager.Application/Controllers/CompetitorsController.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using System.Web.Http;
 using Tournament.Manager.Business.CompetitionConfiguration.CompetitionInfos;
 using Tournament.Manager.Business.DTO;
+using Tournament.Manager.Business.DTO.CompetitionCreation;
 using Tournament.Manager.Business.Services;
 using Tournament.Manager.GridUtils;
 
@@ -27,6 +28,21 @@ namespace Tournament.Manager.Application.Controllers
                     CompetitorColumns = competitorColumns,
                     Competitors = competitors
                 });
+            }
+        }
+
+        [Route("{competitionId}/update")]
+        [HttpPost]
+        public async Task<IHttpActionResult> UpdateCompetitors(int competitionId, [FromBody]List<CompetitorCreationInfoDTO> competitors)
+        {
+            using (var competitionPhaseService = new CompetitionPhaseService())
+            using (var competitorService = new CompetitorService(competitionPhaseService.DbContext))
+            {
+                // TODO: HACK, delete all competition phases only if necessary (player added or removed)
+                competitionPhaseService.DeleteAllCompetitionPhases(competitionId);
+                competitorService.UpdateCompetitors(competitionId, competitors);
+                await competitorService.DbContext.SaveChangesAsync();
+                return Ok();
             }
         }
     }
